requirements/aggregator.txt
numpy
pandas
scikit-learn
psycopg2-binary
sqlalchemy
kafka-python
python-dotenv
requests
numexpr
